refactor: Update ApplyVectorsResponse solved & notSolved field types
